remove a unneeded cast in (int)lun
git-svn-id: svn://anonscm.debian.org/svn/pcsclite/trunk/Drivers/ccid@1046 0ce88b0d-b2fd-0310-8134-9614164e65ea

+++ src/ccid_serial.c
@@ -381,7 +381,7 @@ int ReadChunk(unsigned int lun, unsigned char *buffer, int buffer_length, int mi
 #ifdef DEBUG_LEVEL_COMM
 	char debug_header[] = "<- 121234 ";
 
-	sprintf(debug_header, "<- %06X ", (int)lun);
+	sprintf(debug_header, "<- %06X ", lun);
 #endif
 
 	already_read = 0;
